feat(controller): adds endpoint to display multiple accounts

+++ week-07/day-3/src/main/java/com/greenfoxacademy/bankofsimba/model/BankAccount.java
@@ -12,7 +12,7 @@ public class BankAccount {
   private BankAccount(){
   }
   
-  public BankAccount(String name, int balance, String animalType) {
+  public BankAccount(String name, double balance, String animalType) {
     this();
     this.name = name;
     this.balance = balance;
